refactor ising to binary
